logic/requirements.txt
altgraph==0.17.3
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
importlib-resources==6.0.1
kiwisolver==1.4.5
matplotlib==3.7.2
numpy==1.24.4
packaging==23.1
pefile==2023.2.7
Pillow==10.0.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
pywin32-ctypes==0.2.2
six==1.16.0
zipp==3.16.2
